refactor: move query types to domain.model.query sub-package
- Move BitIndexable, Purity, Sorting, ToplistRange, Category to query sub-package
- Rename WallhavenCategory to Category (package already provides context)
