+++ app/src/main/java/xyz/attacktive/wallhavend/domain/model/query/Sorting.kt
@@ -1,4 +1,4 @@
-package xyz.attacktive.wallhavend.domain.model
+package xyz.attacktive.wallhavend.domain.model.query
 
 import androidx.annotation.StringRes
 import xyz.attacktive.wallhavend.R
