+++ app/src/main/java/xyz/attacktive/wallhavend/domain/repository/WallhavenRepository.kt
@@ -8,9 +8,9 @@ import xyz.attacktive.wallhavend.data.api.dto.WallpaperDto
 import xyz.attacktive.wallhavend.data.api.dto.toDomain
 import xyz.attacktive.wallhavend.domain.model.AppSettings
 import xyz.attacktive.wallhavend.domain.model.NoResultsException
-import xyz.attacktive.wallhavend.domain.model.Sorting
 import xyz.attacktive.wallhavend.domain.model.Wallpaper
-import xyz.attacktive.wallhavend.domain.model.toBitString
+import xyz.attacktive.wallhavend.domain.model.query.Sorting
+import xyz.attacktive.wallhavend.domain.model.query.toBitString
 import xyz.attacktive.wallhavend.domain.service.WallpaperFileManager
 
 @Singleton
